fixes #2
